# src/main/java/com/ericsson/oss/air/csac/controller/datadictionary/DataDictionaryController.java
/*******************************************************************************
 * COPYRIGHT Ericsson 2023
 *
 * The copyright to the computer program(s) herein is the property of
 * Ericsson Inc. The programs may be used and/or copied only with written
 * permission from Ericsson Inc. or in accordance with the terms and
 * conditions stipulated in the agreement/contract under which the
 * program(s) have been supplied.
 ******************************************************************************/

package com.ericsson.oss.air.csac.controller.datadictionary;

import com.ericsson.oss.air.api.DataDictionaryApi;
import com.ericsson.oss.air.api.model.AugmentationListDto;
import com.ericsson.oss.air.api.model.KpiDefinitionListDto;
import com.ericsson.oss.air.api.model.PmDefinitionListDto;
import com.ericsson.oss.air.csac.handler.request.AugmentationRequestHandler;
import com.ericsson.oss.air.csac.handler.request.KPIDefRequestHandler;
import com.ericsson.oss.air.csac.handler.request.PMDefRequestHandler;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RestController;

/**
 * Implementation of DataDictionaryApi generated by openapi generator.
 */
@RestController
public class DataDictionaryController implements DataDictionaryApi {

    @Autowired
    private PMDefRequestHandler pmDefRequestHandler;

    @Autowired
    private KPIDefRequestHandler kpiDefRequestHandler;

    @Autowired
    private AugmentationRequestHandler augmentationRequestHandler;

    @Override
    public ResponseEntity<PmDefinitionListDto> getPmdefs(final Integer start, final Integer rows) {

        final PmDefinitionListDto pmDefinitionListDto = pmDefRequestHandler.getPMDefinitions(start, rows);

        return ResponseEntity.ok().body(pmDefinitionListDto);

    }

    @Override
    public ResponseEntity<KpiDefinitionListDto> getKPIs(final Integer start, final Integer rows) {

        final KpiDefinitionListDto kpiDefinitionListDto = kpiDefRequestHandler.getKPIDefinitions(start, rows);

        return ResponseEntity.ok().body(kpiDefinitionListDto);
    }

    @Override
    public ResponseEntity<AugmentationListDto> getAugmentations(final Integer start, final Integer rows) {

        final AugmentationListDto augmentationListDto = this.augmentationRequestHandler.getAugmentationList(start, rows);

        return ResponseEntity.ok(augmentationListDto);
    }
}