# src/main/resources/v1/eric-oss-core-slice-assurance-cfg-openapi.yaml
...
openapi: "3.0.2"
info:
  x-api-id: 27bdccb8-6e34-4381-9ae7-68dff34800d6
  x-audience: company-internal
  description: |
    This REST API provides the client with the ability to retrieve and browse
    entities previously configured in the data dictionary. Additionally, a client can
    retrieve runtime configuration currently deployed.
  version: "1.0.0"
  title: "Core Slice Assurance Configurator (CSAC)"
  license:
    name: "COPYRIGHT Ericsson 2022"
    url: "https://www.ericsson.com/en/legal"
  termsOfService: "https://www.ericsson.com/en/legal"
...
paths:
  /v1/dictionary/pmdefs:
    get:
      tags:
        - "Data Dictionary"
      summary: "Returns a paged list of PM resources registered in the data dictionary."
      operationId: "GetPmdefs"
      parameters:
        - in: query
          name: start
          description: "The requested start page. Must be greater than or equal to 0."
          schema:
            type: integer
            default: 0
            minimum: 0
        - in: query
          name: rows
          description: "The requested number of rows per page. If a negative number is given, all resources will be returned."
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: "OK. The requested PM resources were successfully retrieved from the data dictionary."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PmDefinitionListDto'
        "400":
          description: "Bad Request. The request cannot be processed due to a client error."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: "Internal Server Error. The server encountered an unrecoverable situation."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "503":
          description: "Service Unavailable. The server cannot handle requests at this time. Please try again later."
  /v1/dictionary/kpis:
    get:
      tags:
        - "Data Dictionary"
      summary: "Returns a paged list of KPI resources registered in the data dictionary."
      operationId: "GetKPIs"
      parameters:
        - in: query
          name: start
          description: "The requested start page. Must be greater than or equal to 0."
          schema:
            type: integer
            default: 0
            minimum: 0
        - in: query
          name: rows
          description: "The requested number of rows per page. If a negative number is given, all resources will be returned."
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: "OK. The requested KPI resources were successfully retrieved from the data dictionary."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KpiDefinitionListDto'
        "400":
          description: "Bad Request. The request cannot be processed due to a client error."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: "Internal Server Error. The server encountered an unrecoverable situation."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "503":
          description: "Service Unavailable. The server cannot handle requests at this time. Please try again later."
  /v1/dictionary/augmentations:
    get:
      tags:         
        - "Data Dictionary"
      summary: "Returns a paged list of augmentation configuration resources registered in the data dictionary."
      operationId: "GetAugmentations"
      parameters:        
        - in: query
          name: start
          description: "The requested start page. Must be greater than or equal to 0."
          schema:             
            type: integer
            default: 0
            minimum: 0
        - in: query
          name: rows
          description:  "The requested number of rows per page. If a negative number is given, all resources will be returned."
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: "OK. The requested resources were successfully retrieved from the data dictionary."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AugmentationListDto'
        "400":
          description: "Bad Request. The request cannot be processed due to a client error."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: "Internal Server Error. The server encountered an unrecoverable situation."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "503":
          description: "Service Unavailable. The server cannot handle requests at this time. Please try again later."
...
components:
  schemas:
    PmDefinitionDto:
      description: "The PM definition schema"
      type: object
      required:
        - name
        - source
      properties:
        name:
          type: string
          description: "The name as it appears in the Avro schema"
          example: "VS_NS_NbrRegisteredSub_5GS"
        source:
          type: string
          description: "The Avro schema reference for Data Catalog lookup"
          example: "5G/PM_COUNTERS/AMF_Mobility_NetworkSlice"
        description:
          type: string
          description: "The description of the PM"
          example: "PCC v1.21:13.1"
        kpis:
          description: "List of all KPI definitions currently referencing this PM definition."
          type: array
          items:
            type: string
          example: [ "RM_RegisteredSubNbrMean_SNSSAI", "RM_RegisteredSubNbrMax_SNSSAI" ]
    PmDefinitionListDto:
      description: "Query result containing the list of PMs as well as information regarding the number of records in this response, total number of records matching the query, and the requested paging information"
      type: object
      required:
        - total
        - count
        - start
        - rows
        - pm_defs
      properties:
        total:
          type: integer
          description: "The total number of PMs matching the query"
          example: 1
        count:
          type: integer
          description: "The number of PMs in this response"
          example: 1
        start:
          type: integer
          description: "The requested start page"
          example: 0
        rows:
          type: integer
          description: "The requested number of rows per page"
          example: 10
        pm_defs:
          description: "The list of PM definitions matching the query for the current page"
          type: array
          items:
            $ref: "#/components/schemas/PmDefinitionDto"
    InputMetricDto:
      description: "The input metrics of a KPI definition"
      type: object
      required:
        - id
        - type
      properties:
        id:
          type: string
          description: "The identifier of a KPI or PM name"
          example: "VS_NS_NbrRegisteredSub_5GS"
        alias:
          type: string
          description: "The alias for local substitution in the KPI's expression. The alias is not shared by multiple KPIs."
          example: "p0"
        type:
          type: string
          description: "Enum type, one of PM_DATA or KPI"
          enum: [ "KPI", "PM_DATA" ]
          example: "PM_DATA"
    KpiDefinitionDto:
      description: "The KPI definition schema"
      type: object
      required:
        - name
        - expression
        - aggregation_type
        - input_metrics
      properties:
        name:
          type: string
          description: "The unique name for the registered KPI"
          example: "slice_registered_users_mean"
        description:
          type: string
          description: "The description of the KPI"
          example: "Mean registered subscribers of network slice through AMF"
        display_name:
          type: string
          description: "The human readable KPI name displayed in the UI"
          example: "Mean Registered Subscribers"
        expression:
          type: string
          description: "The KPI expression in ANSI SQL"
          example: "AVG(p0)"
        aggregation_type:
          type: string
          description: "The aggregation type, (MAX, AVG, etc)"
          example: "AVG"
        aggregation_period:
          type: integer
          description: "If present, indicates the target aggregation period for this KPI definition. This may be overridden in the profile KPI reference"
          example: 15
        is_visible:
          type: boolean
          description: "Determines whether the KPI is consumable by the UI. If false, description and display_name are not used."
          default: true
        input_metrics:
          type: array
          description: "The list of input metrics for this KPI definition"
          minItems: 1
          items:
            $ref: "#/components/schemas/InputMetricDto"
        profiles:
          description: "List of profile definitions referencing this KPI definition."
          type: array
          items:
            type: string
          example: [ "Aggregate based on the SNSSAI for all EC and NF",
                     "Aggregate based on the NF for a specific SNSSAI for all EC",
                     "Aggregate based on the NF for all SNSSAI and EC" ]
    KpiDefinitionListDto:
      description: "The list of all KPI resources registered in the data dictionary for the requested query. Includes also the total number of KPIs matching the query."
      type: object
      required:
        - total
        - count
        - start
        - rows
        - kpi_defs
      properties:
        total:
          type: integer
          description: "The total number of KPIs matching the query"
          example: 1
        count:
          type: integer
          description: "The number of KPIs in this response"
          example: 1
        start:
          type: integer
          description: "The requested start page"
          example: 0
        rows:
          type: integer
          description: "The requested number of rows per page"
          example: 10
        kpi_defs:
          type: array
          description: "List of KPI definitions matching the query"
          items:
            $ref: "#/components/schemas/KpiDefinitionDto"
...
    AugmentationListDto:
      description: "Query result containing the list of augmentation configurations."
      type: object
      properties:
        total:
          type: integer
          description: "The total number of augmentation configurations matching the query"
          example: 1
        count:
          type: integer
          description: "The number of records in this response"
          example: 1
        start:
          type: integer
          description: "The requested start page"
          example: 0
        rows:
          type: integer
          description: "The requested number of rows per page"
          example: 10
        augmentations:
          description: "The list of augmentation definitions matching the query for the current page"
          type: array
          items:
            $ref: "#/components/schemas/AugmentationDto"
    AugmentationFieldDto:
      description: "Augmentation field definition comprising an output field and a list of input fields for the dimensioning query."
      type: object
      properties:
        output:
          type: string
          example: "nsi"
        input:
          type: array
          items:
            type: string
          example: ["snssai","nodeFDN"]
    AugmentationRuleDto:
      description: "Augmentation rule comprising an input schema reference and a list of associated fields"
      type: object
      properties:
        input_schema:
          type: string
          example: "5G|PM_COUNTERS|AMF_Mobility_NetworkSlice_1"
        fields:
          type: array
          items:
            $ref: "#/components/schemas/AugmentationFieldDto"
          example: [
              {
             "output": "nsi",
             "input": [ "snssai", "nodeFDN" ]
           },
           {
              "output": "ec",
              "input": [ "snssai", "nodeFDN" ]
            }
            ]
    AugmentationDto:
      description: "Augmentation configuration instance"
      type: object
      properties:
        ardq_id:
          type: string
          description: "Unique identifer for this augmentation configuration"
          example: "cardq"
        ardq_url:
          type: string
          description: "Target assurance dimensioning query service URL"
          example: "http://eric-oss-cardq:8080"
        ardq_type:
          type: string
          description: "Optional dimensioning query type. The target ARDQ service may or may not require the type qualifier for dimensioning query requests"
          example: "core"
        ardq_rules:          
          description: "List of augmentation rules."
          type: array
          items:
            $ref: '#/components/schemas/AugmentationRuleDto'
        profiles:
          description: "List of associated profile names. At least one entry indicates that this augmentation configuration is currently in effect."
          type: array
          items:
            type: string
          example: [ "Aggregate based on the SNSSAI for all EC and NF",
          "Aggregate based on the NF for a specific SNSSAI for all EC" ]
...
    Problem:
      description: Provides additional information in an error response, RFC 7807.
      type: object
      properties:
        type:
          type: string
          example: "about:blank"
        title:
          type: string
          example: "Method Not Allowed"
        status:
          type: integer
          example: 405
        detail:
          type: string
          example: "This method is not currently supported"
        instance:
          type: string
          example: "/v1/dictionary/pmdefs"